Changed format of String fullyQulifedName in PackageScanner.
ApplicationContainer now checks each class for the @component annotation
If annotation is present, gets the constructors, methods, and fields of that class.
Checks each constructor for the @injection annotation
Gets a class[] of dependency types from the constructors parameters, then gets the dependencies constructors.
Next we need to use this constructor to instantiate the object, and THEN INJECT IT BABY!!!!!